requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
duckdb==1.3.1
et_xmlfile==2.0.0
Flask==3.1.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.2.3
gspread==6.2.1
gspread-dataframe==4.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
narwhals==1.44.0
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.41
streamlit==1.46.1
streamlit-aggrid==1.1.6
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
